@@ -11,7 +11,7 @@ var clearCmd = &cobra.Command{
 	Use:   "clear",
 	Short: "Clear the local database",
 	Long:  `Remove all data and the database file. This action requires confirmation.`,
-	RunE: func(cmd *cobra.Command, args []string) error {
+	RunE: func(cmd *cobra.Command, _ []string) error {
 		ctx := cmd.Context()
 		return runClear(ctx)
 	},
@@ -21,7 +21,7 @@ func init() {
 	clearCmd.Flags().BoolP("force", "f", false, "Skip confirmation prompt")
 }
 
-func runClear(ctx context.Context) error {
+func runClear(_ context.Context) error {
 	// TODO: Implement clear functionality
 	fmt.Println("Clear command not yet implemented")
 	return nil
