@@ -11,7 +11,7 @@ var listCmd = &cobra.Command{
 	Use:   "list",
 	Short: "List all repositories in the local database",
 	Long:  `Display all repositories in the local database with basic information.`,
-	RunE: func(cmd *cobra.Command, args []string) error {
+	RunE: func(cmd *cobra.Command, _ []string) error {
 		ctx := cmd.Context()
 		return runList(ctx)
 	},
@@ -23,7 +23,7 @@ func init() {
 	listCmd.Flags().StringP("format", "f", "table", "Output format (table, json, csv)")
 }
 
-func runList(ctx context.Context) error {
+func runList(_ context.Context) error {
 	// TODO: Implement list functionality
 	fmt.Println("List command not yet implemented")
 	return nil
